+++ HarryPeloteur_DAL/Models/PersonneDto.cs
@@ -12,9 +12,9 @@ namespace HarryPeloteur_DAL
         public int SalleActuelle { get; set; }
         public string Nom { get; set; }
         public int Pv { get; set; }
-        public float Force { get; set; }
-        public float Fuite { get; set; }
-        public float Dexterite { get; set; }
+        public int Force { get; set; }
+        public int Fuite { get; set; }
+        public int Dexterite { get; set; }
         public int Xp { get; set; }
         public int Po { get; set; } 
     }
